requirements.txt
streamlit>=1.40.0
openai>=1.3.8
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
supabase>=2.0.0
